refactor : change folder structure
