+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:wanna_exercise_app/theme/light_theme.dart';
+import 'package:wanna_exercise_app/themes/light_theme.dart';
 import 'package:wanna_exercise_app/UI/pages/home/home_page.dart';
 
 void main() {
