+++ lib/pages/map/widgets/event_card_widget.dart
