+++ lib/pages/map/widgets/search_bar_widget.dart
